+++ docs/src/content/en/reference/auth/workos.mdx
@@ -5,7 +5,7 @@ description: "API reference for the MastraAuthWorkos class, which authenticates 
 
 # MastraAuthWorkos Class
 
-The `MastraAuthWorkos` class provides authentication for Mastra using WorkOS. It verifies incoming requests using WorkOS access tokens and integrates with the Mastra server using the `experimental_auth` option.
+The `MastraAuthWorkos` class provides authentication for Mastra using WorkOS. It verifies incoming requests using WorkOS access tokens and integrates with the Mastra server using the `auth` option.
 
 ## Usage example
 
@@ -16,7 +16,7 @@ import { MastraAuthWorkos } from "@mastra/auth-workos";
 export const mastra = new Mastra({
   // ..
   server: {
-    experimental_auth: new MastraAuthWorkos({
+    auth: new MastraAuthWorkos({
       apiKey: process.env.WORKOS_API_KEY,
       clientId: process.env.WORKOS_CLIENT_ID,
     }),
