+++ examples/experimental-auth-weather-agent/src/mastra/index.ts
@@ -1,15 +1,13 @@
 import { Mastra } from '@mastra/core/mastra';
 
 import { weatherAgent } from './agents';
-import { weatherWorkflow as legacyWeatherWorkflow } from './workflows';
-import { weatherWorkflow } from './workflows/new-workflow';
+import { weatherWorkflow } from './workflows';
 import { authConfig } from './auth';
 
 export const mastra = new Mastra({
   agents: { weatherAgent },
-  legacy_workflows: { legacyWeatherWorkflow },
   workflows: { weatherWorkflow },
   server: {
-    experimental_auth: authConfig,
+    auth: authConfig,
   },
 });
